+++ user/echo.c
@@ -1,11 +1,10 @@
-#include "kernel/types.h"
 #include "user/user.h"
 
 int main(int argc, char* argv[]) {
   int i;
 
   for (i = 1; i < argc; i++) {
-    _write(1, argv[i], strlen(argv[i]));
+    _write(1, argv[i], (int)strlen(argv[i]));
     if (i + 1 < argc) {
       _write(1, " ", 1);
     } else {
